Add WRAP_UNICODE feature
Setting this feature allows breaking lines on all breakpoints considered
legal by Unicode standard, but is incompatible with ASS. This is useful
for non-ASS renderers utilising libass, especially for languanges
typically not utilisng spaces to separate words, but also to correctly
break on punctuation even in languages that do. It does _not_ allow
breaking words apart, so line overflow might still happen.
For this to work a new optional dependency on libunibrak is added.
We assume FriBidi's and libunibreak's utf32/glyph type match.

This is based on an earlier patch by GitHub user siikamiika.